Modified converting RPH value of Time (in the case of outage) to display decimal value of seconds.

+++ yukon-client/report/src/com/cannontech/analysis/tablemodel/MeterOutageModel.java
@@ -229,7 +229,7 @@ public class MeterOutageModel extends ReportModelBase
 				    return meterPD.getTimeStamp();
 
 				case DURATION:
-				    return convertSecondsToTimeString(meterPD.getValue().intValue());
+				    return convertSecondsToTimeString(meterPD.getValue());
 			}
 		}
 		return null;
